Update transport rules

+++ model/social/tasks.go
@@ -4,11 +4,11 @@ import (
 	"math"
 )
 
-func NumBatchesSimple(totalQuantity, transportQuantity int) int {
+func NumBatchesSimple(totalQuantity, transportQuantity uint16) int {
 	return NumBatches(totalQuantity, transportQuantity, transportQuantity)
 }
 
-func NumBatches(totalQuantity, minTransportQuantity, maxTransportQuantity int) int {
+func NumBatches(totalQuantity, minTransportQuantity, maxTransportQuantity uint16) int {
 	if totalQuantity < minTransportQuantity {
 		return 0
 	}
